feat: add basic errors handling

+++ CQRS/src/Application/UserProfiles/CommandHandlers/DeleteUserProfileCommandHandler.cs
@@ -6,7 +6,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.UserProfiles.CommandHandlers;
 
-internal class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand>
+internal class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand, OperationResult<UserProfile>>
 {
     private readonly DataContext _ctx;
 
@@ -15,13 +15,42 @@ internal class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfi
         _ctx = ctx;
     }
 
-    public async Task Handle(DeleteUserProfileCommand request, CancellationToken cancellationToken)
+    public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfileCommand request,
+        CancellationToken cancellationToken)
     {
-        UserProfile? userProfile =
-            await _ctx.UserProfiles.FirstOrDefaultAsync(
-                userProfile => userProfile.UserProfileId == request.UserProfileId, cancellationToken);
+        OperationResult<UserProfile> result = new();
 
-        _ctx.UserProfiles.Remove(userProfile!);
-        await _ctx.SaveChangesAsync(cancellationToken);
+        try
+        {
+            UserProfile? userProfile =
+                await _ctx.UserProfiles.FirstOrDefaultAsync(
+                    userProfile => userProfile.UserProfileId == request.UserProfileId, cancellationToken);
+
+            if (userProfile is null)
+            {
+                result.IsError = true;
+                result.Errors.Add(
+                    new Error
+                    {
+                        Code = ErrorCode.NotFound,
+                        Message = $"No user found with profile Id: {request.UserProfileId}"
+                    }
+                );
+
+                return result;
+            }
+
+            _ctx.UserProfiles.Remove(userProfile!);
+            await _ctx.SaveChangesAsync(cancellationToken);
+        }
+        catch (Exception ex)
+        {
+            result.IsError = true;
+            result.Errors.Add(
+                new Error { Code = ErrorCode.InternalServerError, Message = ex.Message }
+            );
+        }
+
+        return result;
     }
 }
